Cleanup serde_as integration test
Move code into a folder an make use of module.
Split single large files into multiple modules.
Merge the tests serde_as_duration and serde_as_macro into this single
integration test too. They do not need to be standalone tests now that
modules are used.
