+++ tests/serde_as/serde_as_macro.rs
@@ -1,6 +1,4 @@
-use pretty_assertions::assert_eq;
-use serde::{Deserialize, Serialize};
-use serde_with::{serde_as, DisplayFromStr};
+use super::*;
 
 /// Test that the [`serde_as`] macro can replace the `_` type and the resulting code compiles.
 #[test]
@@ -20,46 +18,46 @@ fn test_serde_as_macro_replace_infer_type() {
         e: Box<[u32]>,
     }
 
-    let data = Data {
-        a: 10,
-        b: vec![20, 33],
-        c: vec![(40, "Hello".into()), (55, "World".into()), (60, "!".into())],
-        d: [70, 88],
-        e: vec![99, 100, 110].into_boxed_slice(),
-    };
-    let expected = r##"{
-  "a": 10,
-  "b": [
-    20,
-    33
-  ],
-  "c": [
-    [
-      40,
-      "Hello"
-    ],
-    [
-      55,
-      "World"
-    ],
-    [
-      60,
-      "!"
-    ]
-  ],
-  "d": [
-    70,
-    88
-  ],
-  "e": [
-    99,
-    100,
-    110
-  ]
-}"##;
-
-    assert_eq!(expected, serde_json::to_string_pretty(&data).unwrap());
-    assert_eq!(data, serde_json::from_str(expected).unwrap());
+    is_equal(
+        Data {
+            a: 10,
+            b: vec![20, 33],
+            c: vec![(40, "Hello".into()), (55, "World".into()), (60, "!".into())],
+            d: [70, 88],
+            e: vec![99, 100, 110].into_boxed_slice(),
+        },
+        expect![[r#"
+        {
+          "a": 10,
+          "b": [
+            20,
+            33
+          ],
+          "c": [
+            [
+              40,
+              "Hello"
+            ],
+            [
+              55,
+              "World"
+            ],
+            [
+              60,
+              "!"
+            ]
+          ],
+          "d": [
+            70,
+            88
+          ],
+          "e": [
+            99,
+            100,
+            110
+          ]
+        }"#]],
+    );
 }
 
 /// Test that the [`serde_as`] macro supports `deserialize_as`
@@ -76,12 +74,13 @@ fn test_serde_as_macro_deserialize() {
         c: (u32, u32),
     }
 
-    let data = Data {
-        a: 10,
-        b: vec![20, 33],
-        c: (40, 55),
-    };
-    let expected = r##"{
+    check_deserialization(
+        Data {
+            a: 10,
+            b: vec![20, 33],
+            c: (40, 55),
+        },
+        r##"{
   "a": "10",
   "b": [
     "20",
@@ -91,9 +90,8 @@ fn test_serde_as_macro_deserialize() {
     "40",
     55
   ]
-}"##;
-
-    assert_eq!(data, serde_json::from_str(expected).unwrap());
+}"##,
+    );
 }
 
 /// Test that the [`serde_as`] macro supports `serialize_as`
@@ -110,24 +108,25 @@ fn test_serde_as_macro_serialize() {
         c: (u32, u32),
     }
 
-    let data = Data {
-        a: 10,
-        b: vec![20, 33],
-        c: (40, 55),
-    };
-    let expected = r##"{
-  "a": "10",
-  "b": [
-    "20",
-    "33"
-  ],
-  "c": [
-    "40",
-    55
-  ]
-}"##;
-
-    assert_eq!(expected, serde_json::to_string_pretty(&data).unwrap());
+    check_serialization(
+        Data {
+            a: 10,
+            b: vec![20, 33],
+            c: (40, 55),
+        },
+        expect![[r#"
+        {
+          "a": "10",
+          "b": [
+            "20",
+            "33"
+          ],
+          "c": [
+            "40",
+            55
+          ]
+        }"#]],
+    );
 }
 
 /// Test that the [`serde_as`] macro supports `serialize_as` and `deserialize_as`
@@ -150,25 +149,25 @@ fn test_serde_as_macro_serialize_deserialize() {
         c: (u32, u32),
     }
 
-    let data = Data {
-        a: 10,
-        b: vec![20, 33],
-        c: (40, 55),
-    };
-    let expected = r##"{
-  "a": "10",
-  "b": [
-    "20",
-    "33"
-  ],
-  "c": [
-    "40",
-    55
-  ]
-}"##;
-
-    assert_eq!(expected, serde_json::to_string_pretty(&data).unwrap());
-    assert_eq!(data, serde_json::from_str(expected).unwrap());
+    is_equal(
+        Data {
+            a: 10,
+            b: vec![20, 33],
+            c: (40, 55),
+        },
+        expect![[r#"
+        {
+          "a": "10",
+          "b": [
+            "20",
+            "33"
+          ],
+          "c": [
+            "40",
+            55
+          ]
+        }"#]],
+    );
 }
 
 /// Test that the [`serde_as`] macro works correctly if applied multiple times to a field
@@ -182,11 +181,11 @@ fn test_serde_as_macro_multiple_field_attributes() {
         a: u32,
     }
 
-    let data = Data { a: 10 };
-    let expected = r##"{
-  "a": "10"
-}"##;
-
-    assert_eq!(expected, serde_json::to_string_pretty(&data).unwrap());
-    assert_eq!(data, serde_json::from_str(expected).unwrap());
+    is_equal(
+        Data { a: 10 },
+        expect![[r#"
+        {
+          "a": "10"
+        }"#]],
+    );
 }
